Fix deprecation warnings

+++ tests/test_visualize_matplotlib.py
@@ -390,27 +390,30 @@ class TestVisualizeMatplotlib(unittest.TestCase):
         os.remove("test.pdf")
 
         # Test tikzplotlib
-        viz.save("test.tikz", include_disclaimer=False)
-        self.assertTrue(os.path.exists("test.tikz"))
+        # Note https://github.com/nschloe/tikzplotlib/issues/559
+        matplotlib_version = matplotlib.__version__.split(",")
+        if matplotlib_version[0] == 3 and matplotlib_version[1] < 8:
+            viz.save("test.tikz", include_disclaimer=False)
+            self.assertTrue(os.path.exists("test.tikz"))
 
-        # We save the file size so that we can compare it to the "cleaned"
-        # version
-        file_size_og = os.stat("test.tikz").st_size
+            # We save the file size so that we can compare it to the "cleaned"
+            # version
+            file_size_og = os.stat("test.tikz").st_size
 
-        os.remove("test.tikz")
+            os.remove("test.tikz")
 
-        # Test with clean figure
-        viz.save(
-            "test.tikz",
-            tikz_clean_figure=True,
-            target_resolution=200,
-            include_disclaimer=False,
-        )
-        self.assertTrue(os.path.exists("test.tikz"))
-        # The cleaned file should be smaller
-        file_size_cleaned = os.stat("test.tikz").st_size
-        self.assertLess(file_size_cleaned, file_size_og)
-        os.remove("test.tikz")
+            # Test with clean figure
+            viz.save(
+                "test.tikz",
+                tikz_clean_figure=True,
+                target_resolution=200,
+                include_disclaimer=False,
+            )
+            self.assertTrue(os.path.exists("test.tikz"))
+            # The cleaned file should be smaller
+            file_size_cleaned = os.stat("test.tikz").st_size
+            self.assertLess(file_size_cleaned, file_size_og)
+            os.remove("test.tikz")
 
     def test_save_from_dir_name_ext(self):
         plt.plot([1, 1], [2, 2])
